Adds partition_index
Often times there are algorithms that are better implemented knowing the
midpoint of the partition, rather than the two slices.

partition_index returns the first index that evaluates to false, or one
past the end if all elements evaluated to true.

I updated the tests and benches, and mentioned it in the readme, but I
didn't update the benchmark table in the readme, since I'm on a
different machine than those statistics were originally collected with.